backend/requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0 ; python_version >= '3.8'
anyio==4.3.0 ; python_version >= '3.8'
blinker==1.7.0 ; python_version >= '3.8'
cachelib==0.12.0 ; python_version >= '3.8'
cachetools==5.3.3 ; python_version >= '3.7'
certifi==2024.2.2 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
distro==1.9.0 ; python_version >= '3.6'
flask==3.0.2
flask-session==0.6.0
google-ai-generativelanguage==0.4.0 ; python_version >= '3.7'
google-api-core==2.17.1 ; python_version >= '3.7'
google-auth==2.28.2 ; python_version >= '3.7'
google-generativeai==0.4.1
googleapis-common-protos==1.63.0 ; python_version >= '3.7'
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0 ; python_version >= '3.7'
httpcore==1.0.4 ; python_version >= '3.8'
httpx==0.27.0 ; python_version >= '3.8'
idna==3.6 ; python_version >= '3.5'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.3 ; python_version >= '3.7'
markupsafe==2.1.5 ; python_version >= '3.7'
openai==1.14.1
proto-plus==1.23.0 ; python_version >= '3.6'
protobuf==4.25.3 ; python_version >= '3.8'
pyasn1==0.5.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==2.6.4 ; python_version >= '3.8'
pydantic-core==2.16.3 ; python_version >= '3.8'
python-dotenv==1.0.1
requests==2.31.0 ; python_version >= '3.7'
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
sniffio==1.3.1 ; python_version >= '3.7'
tqdm==4.66.2 ; python_version >= '3.7'
typing-extensions==4.10.0 ; python_version >= '3.8'
urllib3==2.2.1 ; python_version >= '3.8'
werkzeug==3.0.1 ; python_version >= '3.8'
